+++ dtmcli-core/pom.xml
@@ -24,7 +24,7 @@
         <dependency>
             <groupId>io.github.dtm-labs</groupId>
             <artifactId>dtmcli-common</artifactId>
-            <version>2.1.5</version>
+            <version>2.1.8</version>
         </dependency>
 
         <dependency>
@@ -36,6 +36,7 @@
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
         </dependency>
+
     </dependencies>
